requirements.txt
certifi==2019.3.9
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2
django-heroku==0.3.1
gunicorn==19.9.0
idna==2.8
psycopg2==2.8.1
pytz==2018.9
requests==2.21.0
sqlparse==0.3.0
urllib3==1.24.1
whitenoise==4.1.2
atomicwrites==1.3.0
attrs==18.2.0
django-braces==1.13.0
django-cors-headers==2.5.2
django-oauth-toolkit==1.2.0
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.9.2
more-itertools==6.0.0
oauthlib==3.0.1
Pillow==5.4.1
pluggy==0.9.0
py==1.8.0
PyJWT==1.7.1
pytest==4.3.0
python-dateutil==2.8.0
python3-openid==3.1.0
requests-oauthlib==1.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.1.0
defusedxml==0.5.0
export==0.1.2
psycopg2==2.8.1